+++ examples/simple_zonal_with_hub_kubeconfig/hub.tf
@@ -16,7 +16,7 @@
 
 module "hub" {
   source  = "terraform-google-modules/kubernetes-engine/google//modules/hub-legacy"
-  version = "~> 29.0"
+  version = "~> 30.0"
 
   project_id              = var.project_id
   location                = "remote"