+++ modules/beta-public-cluster-update-variant/firewall.tf
@@ -90,7 +90,6 @@ resource "google_compute_firewall" "tpu_egress" {
   ]
 }
 
-
 /******************************************
   Allow GKE master to hit non 443 ports for
   Webhooks/Admission Controllers